backend/requirements.txt
flask==2.0.1
flask-cors==4.0.0
openai-whisper==20231117
numpy==1.24.3
nltk==3.8.1
python-dotenv==1.0.0 